+++ app/src/main/java/com/jacksonasantos/travelplan/dao/interfaces/BrokerIDAO.java
@@ -6,11 +6,11 @@ import java.util.ArrayList;
 import java.util.List;
 
 public interface BrokerIDAO {
-    Broker fetchBrokerById(Long id);
+    Broker fetchBrokerById(Integer id);
     List<Broker> fetchAllBroker();
     ArrayList<Broker> fetchArrayBroker();
     boolean addBroker(Broker broker);
-    void deleteBroker(Long id);
+    void deleteBroker(Integer id);
     boolean updateBroker(Broker broker);
 
 }
